+++ src/orquestra/sdk/_base/_workflow.py
@@ -155,6 +155,7 @@ class WorkflowDef(Generic[_R]):
         project_dir: Optional[Union[str, Path]] = None,
         workspace_id: Optional[WorkspaceId] = None,
         project_id: Optional[ProjectId] = None,
+        dry_run: bool = False,
     ) -> _api.WorkflowRun:
         """
         Schedules workflow for execution.
@@ -167,7 +168,8 @@ class WorkflowDef(Generic[_R]):
                 working directory is used.
             workspace_id: ID of the workspace for workflow - supported only on CE
             project_id: ID of the project for workflow - supported only on CE
-
+            dry_run: Run the workflow without actually executing any task code.
+                Useful for testing infrastructure, dependency imports, etc.
         Raises:
             orquestra.sdk.exceptions.DirtyGitRepo: (warning) when a task def used by
                 this workflow def has a "GitImport" and the git repo that contains it
@@ -184,6 +186,7 @@ class WorkflowDef(Generic[_R]):
             workspace_id=workspace_id,
             project_id=project_id,
             project_dir=project_dir,
+            dry_run=dry_run,
         )
 
     def with_resources(