+++ src/orquestra/sdk/_ray/_build_workflow.py
@@ -20,7 +20,7 @@ from .._base._env import (
 )
 from .._base._logs import _markers
 from ..kubernetes.quantity import parse_quantity
-from ..schema import _compat, ir, responses, workflow_run
+from ..schema import ir, responses, workflow_run
 from . import _client, _id_gen
 from ._client import RayClient
 from ._wf_metadata import InvUserMetadata, pydatic_to_json_dict
@@ -155,6 +155,32 @@ class ArgumentUnwrapper:
         return self._user_fn(*args, **kwargs)
 
 
+SENTINEL = "dry_run task output"
+
+
+def _generate_nop_function(output_metadata: ir.TaskOutputMetadata):
+    def nop(*_, **__):
+        if not output_metadata.is_subscriptable:
+            return SENTINEL
+        else:
+            return (SENTINEL,) * output_metadata.n_outputs
+
+    return nop
+
+
+def _get_user_function(
+    user_fn_ref, dry_run, output_metadata: t.Optional[ir.TaskOutputMetadata]
+):
+    # Ref is None only in case of the data aggregation step
+    if user_fn_ref is None:
+        return _aggregate_outputs
+    elif not dry_run:
+        return _locate_user_fn(user_fn_ref)
+    else:
+        assert output_metadata is not None
+        return _generate_nop_function(output_metadata)
+
+
 def _make_ray_dag_node(
     client: RayClient,
     ray_options: t.Mapping,
@@ -162,9 +188,10 @@ def _make_ray_dag_node(
     ray_kwargs: t.Mapping[str, t.Any],
     args_artifact_nodes: t.Mapping,
     kwargs_artifact_nodes: t.Mapping,
-    n_outputs: t.Optional[int],
     project_dir: t.Optional[Path],
     user_fn_ref: t.Optional[ir.FunctionRef],
+    output_metadata: t.Optional[ir.TaskOutputMetadata],
+    dry_run: bool,
 ) -> _client.FunctionNode:
     """
     Prepares a Ray task that fits a single ir.TaskInvocation. The result is a
@@ -179,10 +206,13 @@ def _make_ray_dag_node(
             see ArgumentUnwrapper
         kwargs_artifact_nodes: a map of keyword arg name to artifact node
             see ArgumentUnwrapper
-        n_outputs: the number of outputs for this task function (if known)
         project_dir: the working directory the workflow was submitted from
         user_fn_ref: function reference for a function to be executed by Ray.
             if None - executes data aggregation step
+        output_metadata: output metadata for the user task function. Keeps number
+            of outputs and if the output is subscriptable
+        dry_run: Run the task without actually executing any user code.
+            Useful for testing infrastructure, dependency imports, etc.
     """
 
     @client.remote
@@ -202,12 +232,10 @@ def _make_ray_dag_node(
                 if project_dir is not None:
                     dispatch.ensure_sys_paths([str(project_dir)])
 
-                if user_fn_ref is None:
-                    serialization = False
-                    user_fn = _aggregate_outputs
-                else:
-                    serialization = True
-                    user_fn = _locate_user_fn(user_fn_ref)
+                # True for all the steps except data aggregation
+                serialization = user_fn_ref is not None
+
+                user_fn = _get_user_function(user_fn_ref, dry_run, output_metadata)
 
                 wrapped = ArgumentUnwrapper(
                     user_fn=user_fn,
@@ -225,14 +253,14 @@ def _make_ray_dag_node(
                 )
                 unpacked: t.Tuple[responses.WorkflowResult, ...]
 
-                if n_outputs is not None and n_outputs > 1:
+                if output_metadata is not None and output_metadata.n_outputs > 1:
                     unpacked = tuple(
                         serde.result_from_artifact(
                             wrapped_return[i], ir.ArtifactFormat.AUTO
                         )
                         if serialization
                         else wrapped_return[i]
-                        for i in range(n_outputs)
+                        for i in range(output_metadata.n_outputs)
                     )
                 else:
                     unpacked = (packed,)
@@ -341,6 +369,7 @@ def make_ray_dag(
     client: RayClient,
     workflow_def: ir.WorkflowDef,
     workflow_run_id: workflow_run.WorkflowRunId,
+    dry_run: bool,
     project_dir: t.Optional[Path] = None,
 ):
     # a mapping of "artifact ID" <-> "the ray Future needed to get the value"
@@ -423,9 +452,10 @@ def make_ray_dag(
             ray_kwargs=kwargs,
             args_artifact_nodes=pos_args_artifact_nodes,
             kwargs_artifact_nodes=kwargs_artifact_nodes,
-            n_outputs=_compat.n_outputs(task_def=user_task, task_inv=invocation),
             project_dir=project_dir,
             user_fn_ref=user_task.fn_ref,
+            output_metadata=user_task.output_metadata,
+            dry_run=dry_run,
         )
 
         for output_id in invocation.output_ids:
@@ -455,9 +485,12 @@ def make_ray_dag(
         ray_kwargs={},
         args_artifact_nodes=pos_args_artifact_nodes,
         kwargs_artifact_nodes={},
-        n_outputs=len(pos_args),
         project_dir=None,
         user_fn_ref=None,
+        output_metadata=ir.TaskOutputMetadata(
+            n_outputs=len(pos_args), is_subscriptable=False
+        ),
+        dry_run=False,
     )
 
     # Data aggregation step is run with catch_exceptions=True - so it returns tuple of