+++ tests/sdk/test_in_process_runtime.py
@@ -77,14 +77,16 @@ def test_secret_inside_ir(
     get_secret.return_value = mocked_secret
     monkeypatch.setattr(_client.SecretsClient, "get_secret", get_secret)
 
-    run_id = runtime.create_workflow_run(wf_with_secrets().model, None)
+    run_id = runtime.create_workflow_run(wf_with_secrets().model, None, dry_run=False)
     result = runtime.get_workflow_run_outputs_non_blocking(run_id)
 
     assert result == (serde.result_from_artifact("Mocked", ir.ArtifactFormat.AUTO),)
 
 
 def test_explicit_n_outputs_single(runtime: InProcessRuntime):
-    run_id = runtime.create_workflow_run(wf_with_explicit_n_outputs().model, None)
+    run_id = runtime.create_workflow_run(
+        wf_with_explicit_n_outputs().model, None, dry_run=False
+    )
     result = runtime.get_workflow_run_outputs_non_blocking(run_id)
 
     assert result == (serde.result_from_artifact(True, ir.ArtifactFormat.AUTO),)
@@ -98,7 +100,7 @@ class TestQueriesAfterRunning:
     @staticmethod
     @pytest.fixture
     def run_id(runtime: InProcessRuntime, wf_def) -> WorkflowRunId:
-        run_id = runtime.create_workflow_run(wf_def, None)
+        run_id = runtime.create_workflow_run(wf_def, None, dry_run=False)
         return run_id
 
     class TestGetWorkflowRunOutputs:
@@ -113,8 +115,8 @@ class TestQueriesAfterRunning:
             wf_def1 = wf_pass_tuple().model
             wf_def2 = wf_pass_callables_from_task().model
 
-            run_id1 = runtime.create_workflow_run(wf_def1, None)
-            run_id2 = runtime.create_workflow_run(wf_def2, None)
+            run_id1 = runtime.create_workflow_run(wf_def1, None, dry_run=False)
+            run_id2 = runtime.create_workflow_run(wf_def2, None, dry_run=False)
 
             outputs1 = runtime.get_workflow_run_outputs_non_blocking(run_id1)
             outputs2 = runtime.get_workflow_run_outputs_non_blocking(run_id2)
@@ -138,7 +140,9 @@ class TestQueriesAfterRunning:
         class TestMultipleTaskOutputs:
             @staticmethod
             def test_some_unused(runtime, wf_def_unused_outputs):
-                run_id = runtime.create_workflow_run(wf_def_unused_outputs, None)
+                run_id = runtime.create_workflow_run(
+                    wf_def_unused_outputs, None, dry_run=False
+                )
 
                 assert runtime.get_available_outputs(run_id) == {
                     "invocation-0-task-two-outputs": serde.result_from_artifact(
@@ -148,7 +152,9 @@ class TestQueriesAfterRunning:
 
             @staticmethod
             def test_all_used(runtime, wf_def_all_used):
-                run_id = runtime.create_workflow_run(wf_def_all_used, None)
+                run_id = runtime.create_workflow_run(
+                    wf_def_all_used, None, dry_run=False
+                )
 
                 assert runtime.get_available_outputs(run_id) == {
                     "invocation-0-task-two-outputs": serde.result_from_artifact(
@@ -161,7 +167,7 @@ class TestStop:
     @staticmethod
     def test_existing_run(runtime, wf_def):
         # Given
-        run_id = runtime.create_workflow_run(wf_def, None)
+        run_id = runtime.create_workflow_run(wf_def, None, dry_run=False)
 
         # When
         runtime.stop_workflow_run(run_id)
@@ -184,8 +190,8 @@ class TestListWorkflowRuns:
     @staticmethod
     def test_happy_path(runtime, wf_def):
         # Given
-        _ = runtime.create_workflow_run(wf_def, None)
-        _ = runtime.create_workflow_run(wf_def, None)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
 
         # When
         wf_runs = runtime.list_workflow_runs()
@@ -196,8 +202,8 @@ class TestListWorkflowRuns:
     @staticmethod
     def test_limit(runtime, wf_def):
         # Given
-        _ = runtime.create_workflow_run(wf_def, None)
-        _ = runtime.create_workflow_run(wf_def, None)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
 
         # When
         wf_runs = runtime.list_workflow_runs(limit=1)
@@ -208,7 +214,7 @@ class TestListWorkflowRuns:
     @staticmethod
     def test_state_running(runtime, wf_def):
         # Given
-        _ = runtime.create_workflow_run(wf_def, None)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
 
         # When
         wf_runs = runtime.list_workflow_runs(state=State.RUNNING)
@@ -220,7 +226,7 @@ class TestListWorkflowRuns:
     @staticmethod
     def test_state_succeeded(runtime, wf_def):
         # Given
-        _ = runtime.create_workflow_run(wf_def, None)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
 
         # When
         wf_runs = runtime.list_workflow_runs(state=State.SUCCEEDED)
@@ -232,8 +238,8 @@ class TestListWorkflowRuns:
     @staticmethod
     def test_max_age(runtime, wf_def):
         # Given
-        run_id = runtime.create_workflow_run(wf_def, None)
-        _ = runtime.create_workflow_run(wf_def, None)
+        run_id = runtime.create_workflow_run(wf_def, None, dry_run=False)
+        _ = runtime.create_workflow_run(wf_def, None, dry_run=False)
         runtime._start_time_store[run_id] = _dates.now() - timedelta(days=1)
 
         # When
@@ -261,5 +267,11 @@ def test_project_raises_warning(runtime, wf_def):
     # Given
     with pytest.warns(expected_warning=exceptions.UnsupportedRuntimeFeature):
         _ = runtime.create_workflow_run(
-            wf_def, project=ProjectRef(workspace_id="", project_id="")
+            wf_def, project=ProjectRef(workspace_id="", project_id=""), dry_run=False
         )
+
+
+def test_dry_run_raises_warning(runtime, wf_def):
+    # Given
+    with pytest.warns(expected_warning=exceptions.UnsupportedRuntimeFeature):
+        _ = runtime.create_workflow_run(wf_def, project=None, dry_run=True)